+++ sdk/clojure/adapter-ring/src/main/starfederation/datastar/clojure/adapter/ring.clj
@@ -1,12 +1,21 @@
 (ns starfederation.datastar.clojure.adapter.ring
   (:require
     [starfederation.datastar.clojure.adapter.ring.impl :as impl]
-    [starfederation.datastar.clojure.api.sse :as sse]))
+    [starfederation.datastar.clojure.adapter.common :as ac]
+    [starfederation.datastar.clojure.api.sse :as sse]
+    [starfederation.datastar.clojure.utils :refer [def-clone]]))
+
+
+(def-clone buffer-size      ac/buffer-size)
+(def-clone hold-write-buff? ac/hold-write-buff?)
+(def-clone gzip?            ac/gzip?)
+(def-clone gzip-buffer-size ac/gzip-buffer-size)
+(def-clone charset          ac/charset)
 
 
 (defn ->sse-response
   "Returns a ring response with status 200, specific SSE headers merged
-  with the provided ones and the body is a sse generator implementing
+  with the provided ones and whose body is a sse generator implementing
   `ring.core.protocols/StreamableResponseBody`.
 
   In sync mode, the connection is closed automatically when the handler is
@@ -18,11 +27,21 @@
   - `:on-open`: Mandatory callback (fn [sse-gen] ...) called when the generator
     is ready to send.
   - `:on-close`: callback (fn [sse-gen] ...) called right after the generator
-    has closed it's connection."
-  [ring-request {:keys [status headers on-open on-close]}]
+    has closed it's connection.
+
+  SSE advanced options:
+  see:
+  - [[buffer-size]]
+  - [[hold-write-buff?]]
+  - [[gzip?]]
+  - [[gzip-buffer-size]]
+  - [[charset]]
+  "
+  [ring-request {:keys [status on-open on-close] :as opts}]
   {:pre [(identity on-open)]}
-  (let [sse-gen (impl/->sse-gen on-open on-close)]
+  (let [sse-gen (impl/->sse-gen on-close)]
     {:status (or status 200)
-     :headers (merge headers (sse/headers ring-request))
-     :body sse-gen}))
+     :headers (ac/headers ring-request opts)
+     :body sse-gen
+     ::impl/opts opts}))