+++ sdk/clojure/src/dev/examples/snippets.clj
@@ -15,6 +15,7 @@
 (d*/merge-signals! sse "{prize: '...'}")
 
 ;; setup
+#_{:clj-kondo/ignore true}
 (comment
   (require
     '[starfederation.datastar.clojure.api :as d*]